Initial
- Split `nnom_layer.c`, `nnom_activations.c`, `nnom_run.c` into layer-based files.
- `nnom_out_shape.c` and all `*_out_shape()` now changed to `*_build()`  and placed in each layer-based files.
