+++ src/core/nnom_utils.c
