+++ src/layers/nnom_baselayer.c
@@ -1,0 +1,74 @@
+/*
+ * Copyright (c) 2018-2019
+ * Jianjia Ma, Wearable Bio-Robotics Group (WBR)
+ * majianjia@live.com
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Change Logs:
+ * Date           Author       Notes
+ * 2019-07-23     Jianjia Ma   The first version
+ */
+
+#include <stdint.h>
+#include <string.h>
+#include <stdbool.h>
+
+#include "nnom.h"
+#include "nnom_local.h"
+#include "nnom_layers.h"
+
+// this layer copys the input to the output
+
+nnom_status_t default_build(nnom_layer_t *layer);
+nnom_status_t default_run(nnom_layer_t *layer);
+
+nnom_layer_t *BaseLayer()
+{
+	nnom_io_layer_t *layer;
+	nnom_layer_io_t *in, *out;
+
+	// apply a block memory for all the sub handles.
+	size_t mem_size = sizeof(nnom_io_layer_t) + sizeof(nnom_layer_io_t) * 2;
+	layer = nnom_mem(mem_size);
+	if (layer == NULL)
+		return NULL;
+
+	// distribut the memory to sub handles.
+	in = (void *)((uint8_t*)layer + sizeof(nnom_io_layer_t));
+	out = (void *)((uint8_t*)in + sizeof(nnom_layer_io_t));
+
+	// set type in layer parent
+	layer->super.type = NNOM_BASE;
+	layer->super.run = default_run;
+	layer->super.build = default_build;
+	// set buf state
+	in->type = LAYER_BUF_TEMP;
+	out->type = LAYER_BUF_NULL;
+	// put in & out on the layer.
+	layer->super.in = io_init(layer, in);
+	layer->super.out = io_init(layer, out);
+
+	return (nnom_layer_t *)layer;
+}
+
+// this is call while output shape is not defined.
+// this will set the output shape same as input shape, and it set only the primary IO
+// this cannot be used as first layer, of course...
+nnom_status_t default_build(nnom_layer_t *layer)
+{
+	// get the last layer's output as input shape
+	layer->in->shape = layer->in->hook.io->shape;
+	// output shape
+	layer->out->shape = layer->in->shape;
+
+	return NN_SUCCESS;
+}
+
+// simply copy input to output
+nnom_status_t default_run(nnom_layer_t *layer)
+{
+	if(layer->out->type != LAYER_BUF_NULL)
+		memcpy(layer->out->mem->blk, layer->in->mem->blk, shape_size(&layer->in->shape)); 
+	return NN_SUCCESS;
+}
