+++ src/layers/nnom_concat.c
@@ -1,0 +1,225 @@
+/*
+ * Copyright (c) 2018-2019
+ * Jianjia Ma, Wearable Bio-Robotics Group (WBR)
+ * majianjia@live.com
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Change Logs:
+ * Date           Author       Notes
+ * 2019-07-23     Jianjia Ma   The first version
+ */
+
+#include <stdint.h>
+#include <string.h>
+#include <stdbool.h>
+
+#include "nnom.h"
+#include "nnom_local.h"
+#include "nnom_layers.h"
+
+nnom_status_t concat_build(nnom_layer_t *layer);
+nnom_status_t concat_run(nnom_layer_t *layer);
+
+
+// concate method
+// concate requires more than one input module. aux input will be allocated in model.merge()
+nnom_layer_t *Concat(int8_t axis)
+{
+	nnom_concat_layer_t *layer;
+	nnom_layer_io_t *in, *out;
+	size_t mem_size;
+
+	// apply a block memory for all the sub handles.
+	mem_size = sizeof(nnom_concat_layer_t) + sizeof(nnom_layer_io_t) * 2;
+	layer = nnom_mem(mem_size);
+	if (layer == NULL)
+		return NULL;
+
+	// distribut the memory to sub handles.
+	in = (void *)((uint8_t*)layer + sizeof(nnom_concat_layer_t));
+	out = (void *)((uint8_t*)in + sizeof(nnom_layer_io_t));
+
+	// set type in layer parent
+	layer->super.type = NNOM_CONCAT;
+	layer->super.run = concat_run;
+	layer->super.build = concat_build;
+	// set buf state
+	in->type = LAYER_BUF_TEMP;
+	out->type = LAYER_BUF_TEMP;
+	// put in & out on the layer.
+	layer->super.in = io_init(layer, in);
+	layer->super.out = io_init(layer, out);
+
+	// check the axis
+	{
+		uint32_t shape_element_num = sizeof(nnom_shape_t) / sizeof(nnom_shape_data_t);
+		if (axis < 0)
+			axis = (shape_element_num + axis);
+		else if (axis >0)
+			axis = axis -1; // keras use axis start from 1. we are using 0, 1, 2 (check?)
+		layer->axis = axis; 
+	}
+
+	return (nnom_layer_t *)layer;
+}
+
+
+nnom_status_t concat_build(nnom_layer_t *layer)
+{
+	nnom_concat_layer_t *cl = (nnom_concat_layer_t *)layer;
+	nnom_layer_io_t *in;
+	uint32_t in_num = 0;
+	uint32_t offset;
+	int32_t shape_element_num;
+
+	// for each input module, copy the shape from the output of last layer
+	in = layer->in;
+	while (in != NULL)
+	{
+		//get the last layer's output as input shape
+		in->shape = in->hook.io->shape;
+		in = in->aux;
+		in_num++;
+	}
+
+	// get how many element in shape
+	shape_element_num = sizeof(nnom_shape_t) / sizeof(nnom_shape_data_t);
+	if (cl->axis >= shape_element_num || cl->axis <= -shape_element_num)
+		return NN_ARGUMENT_ERROR;
+
+	// find the axis to concat
+	offset = cl->axis;
+
+	// do the work
+	for (uint32_t i = 0; i < shape_element_num * sizeof(nnom_shape_data_t); i += sizeof(nnom_shape_data_t))
+	{
+		// exclue the concat axies
+		if (i == offset * sizeof(nnom_shape_data_t))
+		{
+			nnom_shape_data_t *out_axis = (nnom_shape_data_t *)((uint8_t*)(&layer->out->shape) + i);
+			*out_axis = 0;
+
+			in = layer->in;
+			while (in != NULL)
+			{
+				*out_axis += *(nnom_shape_data_t *)((uint8_t*)(&in->shape) + i);
+				in = in->aux;
+			}
+			continue;
+		}
+
+		// check others, all other must be same shape
+		in = layer->in;
+		while (in != NULL && in->aux != NULL)
+		{
+			if (*(nnom_shape_data_t *)((uint8_t*)(&in->shape) + i) !=
+				*(nnom_shape_data_t *)((uint8_t*)(&in->aux->shape) + i))
+				return NN_ARGUMENT_ERROR;
+			in = in->aux;
+		}
+
+		// now set other axis
+		*(nnom_shape_data_t *)((uint8_t*)(&layer->out->shape) + i) =
+			*(nnom_shape_data_t *)((uint8_t*)(&layer->in->shape) + i);
+	}
+
+	return NN_SUCCESS;
+}
+
+
+#ifdef NNOM_USING_CHW
+// axis index converter between HWC and CHW
+static inline int chw_i(int hwc)
+{
+	hwc = hwc + 1;			
+	if(hwc>2) 
+		hwc = 0;
+	return hwc;
+}
+static inline int hwc_i(int chw)
+{
+	chw = chw - 1;			
+	if(chw<0) 
+		chw = 2;
+	return chw;
+}
+#endif
+
+nnom_status_t concat_run(nnom_layer_t *layer)
+{
+	// by default, concat layer has mutiple (>=2) input and 1 output.
+	nnom_concat_layer_t *cl = (nnom_concat_layer_t *)layer;
+	nnom_shape_axis_t *in_shape;
+	nnom_layer_io_t *in;
+
+#ifdef NNOM_USING_CHW
+	// Concatenate for HWC	
+	uint8_t *pin;
+	uint8_t *pout = layer->out->mem->blk;
+	uint32_t block_size;
+	uint32_t n_block;
+	
+	// calcualte number of block to concat. the other shapes before the concat axis
+	in_shape = (nnom_shape_axis_t*)(&layer->in->shape);
+	n_block = 1;
+	for(int i= 0; i< chw_i(cl->axis); i++)
+	{
+		n_block *= in_shape->axis[hwc_i(i)];
+	}
+	
+	// concat all input layers
+	for(int i=0; i<n_block; i++)
+	{
+		in = layer->in;
+		while (in != NULL)
+		{
+			in_shape = (nnom_shape_axis_t*)(&in->shape);
+			// the block size of concat data in this layer
+			block_size = 1;
+			for(int j= 2; j >= chw_i(cl->axis); j--)
+				block_size *= in_shape->axis[hwc_i(j)];
+			// concat		
+			pin = (uint8_t *)in->mem->blk + i * block_size;
+			memcpy(pout, pin, block_size);
+			pout += block_size;
+			in = in->aux;
+		}
+	}
+	
+#else // end of HWC concate
+	
+	// Concatenate for HWC	
+	uint8_t *pin;
+	uint8_t *pout = layer->out->mem->blk;
+	uint32_t block_size;
+	uint32_t n_block;
+	
+	// calcualte number of block to concat. the other shapes before the concat axis
+	in_shape = (nnom_shape_axis_t*)(&layer->in->shape);
+	n_block = 1;
+	for(int i=0; i< cl->axis; i++)
+		n_block *= in_shape->axis[i];
+	
+	// concat all input layers
+	for(int i=0; i<n_block; i++)
+	{
+		in = layer->in;
+		while (in != NULL)
+		{
+			in_shape = (nnom_shape_axis_t*)(&in->shape);
+			// the block size of concat data in this layer
+			block_size = 1;
+			for(int j=cl->axis; j < 3; j++)
+				block_size *= in_shape->axis[j];
+			// concat		
+			pin = (uint8_t *)in->mem->blk + i * block_size;
+			memcpy(pout, pin, block_size);
+			pout += block_size;
+			in = in->aux;
+		}
+	}
+#endif
+	return NN_SUCCESS;
+}
+
